ponybot/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-celery-beat==2.2.1
django-configurations==2.2
django-extensions==3.1.3
django-timezone-field==4.1.2
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.9.1
inflection==0.5.1
iniconfig==1.1.1
kombu==5.1.0
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
